requirements.txt
# Poke Bowl Inventory System - Python Dependencies
# For NVIDIA Jetson Orin Nano with JetPack 6.x (Ubuntu 22.04)

# Core computer vision and deep learning
ultralytics>=8.0.0          # YOLO models
opencv-python>=4.8.0        # Computer vision operations (use opencv-python-headless for headless systems)
torch>=2.0.0                # PyTorch (install Jetson-optimized version separately)
torchvision>=0.15.0         # PyTorch vision utilities
numpy>=1.24.0,<2.0.0        # Array operations

# Web server and async operations
aiohttp>=3.9.0              # Async HTTP server and WebSockets
aiofiles>=23.0.0            # Async file operations

# Configuration and data handling
PyYAML>=6.0                 # YAML configuration files
python-dotenv>=1.0.0        # Environment variable management (optional)

# Logging and monitoring (optional but recommended)
colorlog>=6.7.0             # Colored logging output

# Timezone support
pytz>=2023.3                # Timezone conversion for EST timestamps

# ========================================
# JETSON-SPECIFIC INSTALLATION NOTES
# ========================================
#
# 1. PyTorch Installation:
#    The default PyTorch from pip may not work on Jetson.
#    Install the official NVIDIA PyTorch wheel for JetPack 6.x:
#    
#    wget https://developer.download.nvidia.com/compute/redist/jp/v60/pytorch/torch-2.1.0a0+41361538.nv23.06-cp310-cp310-linux_aarch64.whl
#    pip3 install torch-2.1.0a0+41361538.nv23.06-cp310-cp310-linux_aarch64.whl
#
# 2. Torchvision Installation:
#    Install matching torchvision version:
#    
#    sudo apt-get install libjpeg-dev zlib1g-dev
#    git clone --branch v0.16.0 https://github.com/pytorch/vision torchvision
#    cd torchvision
#    python3 setup.py install --user
#
# 3. OpenCV Installation:
#    JetPack includes optimized OpenCV. If you need to install:
#    
#    pip3 install opencv-python-headless  # For headless (no GUI)
#    # OR use the system OpenCV that comes with JetPack
#
# 4. Ultralytics YOLO:
#    Install normally with pip:
#    
#    pip3 install ultralytics
#
# ========================================
# INSTALLATION COMMAND
# ========================================
#
# Standard installation (after installing PyTorch/Torchvision manually):
#   pip3 install -r requirements.txt
#
# For development (with additional tools):
#   pip3 install -r requirements.txt pytest black flake8
#
# ========================================
# SYSTEM DEPENDENCIES (install with apt)
# ========================================
#
# sudo apt-get update
# sudo apt-get install -y \
#     python3-pip \
#     python3-dev \
#     libopencv-dev \
#     python3-opencv \
#     v4l-utils \
#     chromium-browser \
#     gstreamer1.0-tools \
#     gstreamer1.0-plugins-good \
#     gstreamer1.0-plugins-bad
#
# ========================================

